services/auth_api/requirements-auth.txt
fastapi~=0.83.0
pydantic~=1.9.0
aiohttp~=3.8.1
uvicorn~=0.18.3
motor~=3.0.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
sqlalchemy==1.4.42
psycopg2-binary==2.9.5
python-multipart==0.0.5